requirements.txt
requests
yt-dlp
loguru
pyyaml